Introduce the concept of a ETS backed pool of UDP ports
Current implementation uses a single gen server to route every published
metric.
With this new one, the genserver does not route them anymore, and there
is a pool of sockets that can do the work of shipping the metrics.

My original approach at this was to follow the same approach as
https://andrealeopardi.com/posts/process-pools-with-elixirs-registry/
But I realized that I was introducing sweeping changes, completely
changing the suprevision tree and most of the wiring.

So i went back to the drawing board, and tried to accomplish the same
without all that. I did that using a named :ets table
Clients reading can directly go to the :ets.
In case of errors, I'm still using the genserver to serialize the
requests so the port can be re-created a single time

As things stand, this change is backwards compatible with the existing
implementation.

Closes #33