Always report sum as DataDog counter
Previously, sum updates were reported as gauge increments/decrements.
However, this sort of metric updates do not allow to view the sum of
measurements in the DataDog dashboard, because with gauges, only the last
measurement in the flush interval is shown.

With counter, it is possible to view the total sum of measurement values
in a flush interval. To make sure that the actual sum and not a rate is
displayed in the DataDog dashboard, the `as_count()` (https://docs.datadoghq.com/developers/metrics/type_modifiers/?tab=rate#in-application-modifiers)
modifier should be used.