Always report sum as a counter with DataDog formatter (#47)
Previously, sum updates were reported as gauge increments/decrements. However, 
this sort of metric updates do not allow to view the sum of measurements in the 
DataDog dashboard, because with gauges, only the last measurement in the flush 
interval is actually stored in DataDog.

With counter, it is possible to view the total sum of measurement values in a 
flush interval. To make sure that the actual sum and not a rate is displayed in 
the DataDog dashboard, the as_count()
(https://github.com/beam-telemetry/telemetry_metrics_statsd/pull/docs.datadoghq.com/developers/metrics/type_modifiers/?tab=rate#in-application-modifiers) 
modifier should be used.

While this is a breaking change, the previous behaviour was simply wrong.