[feat] Emitter Pools
During testing at Whatnot, we found that when emitting around 50k
metrics per second, the current pool implementation caused the calling
processes to stall while waiting for `:gen_udp` to write data. This
stalling caused significant degradation to our application, and
eventually caused an outage.

To address this, we came up with the concept of using a pool of
emitters. An emitter is a process that contains a socket (either UDP
or domain), receives metrics and writes them to a socket. This
architecture frees up the calling processes so they can write a metric
and not wait on a result. Emitters also allow us to combine metrics so
they fill up a UDP packet. The end result, according to our testing,
moved us from 50k to 780k metrics per second.